🪲 [Fix]: `Disconnect-GithubAccount` with `Context` and `Silent` parameters (#166)
## Description

- `Disconnect-GithubAccount` with `Context` and `Silent` parameters
  - Fixes #165 

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas

+++ src/functions/public/Auth/Disconnect-GitHubAccount.ps1
@@ -36,18 +36,25 @@
         # The context to log out of.
         [Parameter()]
         [Alias('Name')]
-        [string] $Context = (Get-GitHubConfig -Name 'DefaultContext')
+        [string] $Context = (Get-GitHubConfig -Name 'DefaultContext'),
+
+        # Silently disconnects from GitHub.
+        [switch] $Silent
     )
 
     $commandName = $MyInvocation.MyCommand.Name
     Write-Verbose "[$commandName] - Start"
 
-    $Context = Get-GitHubConfig -Name 'DefaultContext'
     Remove-GitHubContext -Context $Context
-    Remove-GitHubConfig -Name 'DefaultContext'
+    $isDefaultContext = $Context -eq (Get-GitHubConfig -Name 'DefaultContext')
+    if ($isDefaultContext) {
+        Remove-GitHubConfig -Name 'DefaultContext'
+    }
 
-    Write-Host '✓ ' -ForegroundColor Green -NoNewline
-    Write-Host "Logged out of GitHub! [$Context]"
+    if (-not $Silent) {
+        Write-Host '✓ ' -ForegroundColor Green -NoNewline
+        Write-Host "Logged out of GitHub! [$Context]"
+    }
 
     Write-Verbose "[$commandName] - End"
 }
